pulumi/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool]
[tool.poetry]
package-mode = false
[tool.poetry.dependencies]
pulumi = "^3.138"
pulumi-aws = "^6.58"
python = "^3.9"
black = "^24.10.0"
flake8 = "^7.1.1"
pre-commit = "^4.0.1"
python-dotenv = "^1.0.1"
